split data_models.go
