+++ persistence/data_models/backoff_immediate_task_req.go
@@ -1,0 +1,12 @@
+// Copyright (c) 2023 xCherryIO Organization
+// SPDX-License-Identifier: Apache-2.0
+
+package data_models
+
+type BackoffImmediateTaskRequest struct {
+	LastFailureStatus    int32
+	LastFailureDetails   string
+	Prep                 PrepareStateExecutionResponse
+	Task                 ImmediateTask
+	FireTimestampSeconds int64
+}
