+++ persistence/data_models/partition_key.go
@@ -1,0 +1,9 @@
+// Copyright (c) 2023 xCherryIO Organization
+// SPDX-License-Identifier: Apache-2.0
+
+package data_models
+
+type PartitionKey struct {
+	Namespace string
+	ProcessId string
+}
